TestPush

+++ ewm-service/src/main/java/ru/practicum/dto/EventRequestStatusUpdateRequest.java
@@ -13,6 +13,6 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EventRequestStatusUpdateRequest {
-    private List<Integer> requestIds;
+    private List<Long> requestIds;
     private EventStatus status;
 }
